Crate hashed_parachain_runtime

Modules

api

constants

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

xcm_config

Structs

AnnouncementDepositBase

AnnouncementDepositFactor

ApprovalDeposit

AssetDeposit

BasicDeposit

BountyCuratorDeposit

BountyDepositBase

BountyDepositPayoutDelay

BountyUpdatePeriod

BountyValueMinimum

Burn

CandidateDeposit

ChallengePeriod

ChildBountyValueMinimum

ChildMaxLen

CollectionDeposit

ConfigDepositBase

CouncilMaxMembers

CouncilMaxProposals

CouncilMotionDuration

CuratorDepositMax

CuratorDepositMin

CuratorDepositMultiplier

CurrencyAdapter

Implements the transaction payment for a pallet implementing the Currency trait (eg. the pallet_balances) using an unbalance handler (implementing OnUnbalanced).

DataDepositPerByte

DepositBase

DepositFactor

DocDescMaxLen

DocDescMinLen

DocNameMaxLen

DocNameMinLen

ExecutiveBody

ExistentialDeposit

FieldDeposit

FriendDepositFactor

GenesisConfig

IndexDeposit

ItemDeposit

KeyLimit

LabelMaxLen

MaxAdditionalFields

MaxApplicants

MaxApplicationsPerCustodian

MaxApprovals

MaxAuthorities

MaxAuthsPerMarket

MaxCandidateIntake

MaxCandidates

MaxCosignersPerVault

MaxFeedbackLen

MaxFiles

MaxFriends

MaxInvulnerables

MaxKeys

MaxLockDuration

MaxLocks

MaxMarketsPerItem

MaxOffersPerMarket

MaxOwnedDocs

MaxParentsInCollection

MaxPeerDataEncodingSize

MaxPeerInHeartbeats

MaxPermissionsPerRole

MaxProposalsPerVault

MaxRegistrars

MaxReserves

MaxRolesPerAuth

MaxRolesPerPallet

MaxRolesPerUser

MaxScopesPerPallet

MaxSharedFromDocs

MaxSharedToDocs

MaxStrikes

MaxSubAccounts

MaxUsersPerRole

MaxVaultsPerUser

MaximumReasonLength

MembershipMaxMembers

MetadataDepositBase

MetadataDepositPerByte

MinCandidates

MinVestedTransfer

MinimumPeriod

NameMaxLen

NotesMaxLen

Offset

OperationalFeeMultiplier

OutputDescriptorMaxLen

PSBTMaxLen

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Percent

A fixed point representation of a number in the range [0, 1].

Period

PeriodSpend

Permill

A fixed point representation of a number in the range [0, 1].

PermissionMaxLen

PotId

PreimageBaseDeposit

PreimageByteDeposit

PreimageMaxSize

ProposalBond

ProposalBondMaximum

ProposalBondMinimum

ProxyDepositBase

ProxyDepositFactor

RecoveryDeposit

ReservedDmpWeight

ReservedXcmpWeight

RoleMaxLen

RotationPeriod

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

SessionLength

SocietyPalletId

SpendPeriod

StringLimit

SubAccountDeposit

TargetedFeeAdjustment

A struct to update the weight multiplier per block. It implements Convert<Multiplier, Multiplier>, meaning that it can convert the previous multiplier to the next one. This should be called on on_finalize of a block, prior to potentially cleaning the weight data from the system pallet.

TipCountdown

TipFindersFee

TipReportDepositBase

TransactionByteFee

Relay Chain TransactionByteFee / 10

TreasuryPalletId

UncleGenerations

UnvestedFundsAllowedWithdrawReasons

ValueLimit

VaultDescriptionMaxLen

Version

WeightToFee

Handles converting a weight scalar to a fee value, based on the scale and granularity of the node’s balance type.

WrongSideDeduction

XPubLen

Enums

MultiAddress

A multi-format address wrapper for on-chain accounts.

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

RuntimeEvent

Constants

DAYS

EXISTENTIAL_DEPOSIT

The existential deposit. Set to 1/10 of the Connected Relay Chain.

HOURS

MICROUNIT

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MILLIUNIT

MINUTES

SLOT_DURATION

UNIT

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Assets

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Authorship

Balance

Balance of an account.

Balances

BalancesConfig

BitcoinVaults

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

Bounties

CheckedExtrinsic

Extrinsic type that has already been checked.

ChildBounties

CollatorSelection

CollatorSelectionConfig

CollatorSelectionUpdateOrigin

ConfidentialDocs

Council

CouncilConfig

CumulusXcm

DmpQueue

Executive

Executive: handles dispatch to the various modules.

Fruniques

GatedMarketplace

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Identity

Index

Index of a transaction in the chain.

Indices

Membership

Multiplier

Fee multiplier.

Multisig

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

Preimage

Proxy

RBAC

RandomnessCollectiveFlip

Recovery

Remark

RootOrThreeFifthsOfCouncil

Session

SessionConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Society

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

Treasury

TreasuryConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Uniques

Vesting

Whitelist

XcmpQueue