Type Definition hashed_parachain_runtime::AllPalletsWithoutSystemReversed

pub type AllPalletsWithoutSystemReversed = (MappedAssets, TemplateModule, FundAdmin, ConfidentialDocs, RBAC, GatedMarketplace, Fruniques, Uniques, BitcoinVaults, Whitelist, Preimage, Remark, Proxy, Sudo, Vesting, Multisig, Membership, Indices, Recovery, Identity, Society, RandomnessCollectiveFlip, Council, Assets, ChildBounties, Treasury, Bounties, DmpQueue, CumulusXcm, PolkadotXcm, XcmpQueue, AuraExt, Aura, Session, CollatorSelection, Authorship, TransactionPayment, Balances, ParachainInfo, Timestamp, ParachainSystem);

👎Deprecated: Using reverse pallet orders is deprecated. use only AllPalletsWithSystem or AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.