Module hashed_parachain_runtime::xcm_config

Structs

DenyReserveTransferToRelayChain

DenyThenTry

Deny executing the xcm message if it matches any of the Deny filter regardless of anything else. If it passes the Deny, and matches one of the Allow cases then it is let through.

MaxAssetsIntoHolding

MaxInstructions

ParentOrParentsExecutivePlurality

RelayChainOrigin

RelayLocation

RelayNetwork

UnitWeightCost

UniversalLocation

XcmConfig

Type Definitions

Barrier

LocalAssetTransactor

Means for transacting assets on this chain.

LocalOriginToLocation

No local origins on this chain are allowed to dispatch XCM sends/executions.

LocationToAccountId

Type for specifying how a MultiLocation can be converted into an AccountId. This is used when determining ownership of accounts for asset transacting and when attempting to use XCM Transact in order to determine the dispatch Origin.

XcmOriginToTransactDispatchOrigin

This is the type we use to convert an (incoming) XCM origin into a local Origin instance, ready for dispatching a transaction with Xcm’s Transact. There is an OriginKind which can biases the kind of local Origin it will become.

XcmRouter

The means for routing XCM messages which are not for local execution into the right message queues.